# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "description": "Example info",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a user",
                "description": "Registers new user",
                "operationId": "c0eaa9600f14ac767e20818010954c59",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Full name",
                                        "type": "string",
                                        "example": "Test"
                                    },
                                    "email": {
                                        "description": "Email address",
                                        "type": "string",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User Created Successfully"
                                        },
                                        "file_name": {
                                            "type": "string",
                                            "example": "9|iwvR6QMeX1s8a83wZFhxvApSbgu7Ee8CJs9N7nR4b05f26e3"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login a user",
                "description": "Login a user",
                "operationId": "e45287a3dc02f4633476dbf218bb9ef7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email address",
                                        "type": "string",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User Logged In Successfully"
                                        },
                                        "file_name": {
                                            "type": "string",
                                            "example": "9|iwvR6QMeX1s8a83wZFhxvApSbgu7Ee8CJs9N7nR4b05f26e3"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/experiments": {
            "get": {
                "tags": [
                    "Experiments"
                ],
                "summary": "Get a list of experiments",
                "description": "Returns all experiments",
                "operationId": "a0534edd21fd0b243acea4da324c8610",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "experiments": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "file_name": {
                                                        "type": "string",
                                                        "example": "experiment_files/1622619815_1619954846_tcn.zcos"
                                                    },
                                                    "context": {
                                                        "type": "string",
                                                        "example": "{}"
                                                    },
                                                    "output": {
                                                        "type": "string",
                                                        "example": "{}"
                                                    },
                                                    "save": {
                                                        "type": "integer",
                                                        "example": 0
                                                    },
                                                    "created_by": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Experiments"
                ],
                "summary": "Create a new experiment",
                "description": "Stores a new experiment",
                "operationId": "c313fedf97b93b728e83282661ec8f70",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "file",
                                    "context",
                                    "output",
                                    "save"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Experiment file",
                                        "type": "string",
                                        "format": "file"
                                    },
                                    "context": {
                                        "description": "Context",
                                        "type": "string"
                                    },
                                    "output": {
                                        "description": "Output",
                                        "type": "string"
                                    },
                                    "save": {
                                        "description": "Save",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Experiment created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/experiments/{id}": {
            "get": {
                "tags": [
                    "Experiments"
                ],
                "summary": "Get an experiment by ID",
                "description": "Returns an experiment based on ID",
                "operationId": "dcb8183b23f8500f17e3f9fec88e1feb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the experiment",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "file_name": {
                                            "type": "string",
                                            "example": "experiment_files/1622619815_1619954846_tcn.zcos"
                                        },
                                        "context": {
                                            "type": "string",
                                            "example": "{}"
                                        },
                                        "output": {
                                            "type": "string",
                                            "example": "{}"
                                        },
                                        "save": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "created_by": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Experiment not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Experiments"
                ],
                "summary": "Delete an experiment by ID",
                "description": "Deletes an experiment based on ID",
                "operationId": "c7759017f9e8eea4653a3adef4c71a29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the experiment",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Experiment deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Experiment deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Experiment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error deleting experiment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "API Endpoints for authentication"
        }
    ]
}